+++ src/lib/handlers/externals/externals.contract.ts
@@ -1,0 +1,10 @@
+import type {SharedInfo, ShareObject, ShareConfig, ShareOptions} from '@softarc/native-federation-runtime';
+
+import type { Version } from '../version/version.contract';
+
+type ExternalsHandler = {
+    fromStorage(scope: string|'global'): Record<string, Version>;
+    toStorage(externals: SharedInfo[], scopeUrl: string): SharedInfo[]
+}
+
+export { ExternalsHandler, SharedInfo, ShareObject, ShareConfig, ShareOptions }