+++ src/lib/handlers/import-map/import-map.contract.ts
@@ -1,13 +1,12 @@
 import type { ImportMap } from '@softarc/native-federation-runtime';
 
-import type { Remote } from '../remote-info/remote-info.contract';
-
+import type { RemoteName } from '../remote-info';
 
 type ImportMapHandler = {
-    toImportMap: (remoteInfo: Remote, remoteName?: string) => ImportMap,
-    createEmpty: () => ImportMap,
-    merge: (maps: ImportMap[]) => ImportMap
+    create: (from?: ImportMap) => ImportMap,
+    fromStorage: (manifest: RemoteName[]) => ImportMap,
+    addToDOM: (map: ImportMap) => ImportMap
 }
 
-export { Imports } from '@softarc/native-federation-runtime';
+export { Imports, Scopes } from '@softarc/native-federation-runtime';
 export { ImportMapHandler, ImportMap}