+++ src/lib/handlers/shared-info/shared-info.contract.ts
@@ -1,9 +1,0 @@
-import type { Remote } from "./../remote-info/remote-info.contract"
-
-type SharedInfoHandler = {
-    mapSharedDeps: (remoteInfo: Remote) => Record<string, string>,
-    addToCache: (remoteInfo: Remote) => Remote
-}
-
-export { SharedInfoHandler }
-export type { SharedInfo } from '@softarc/native-federation-runtime';