+++ src/lib/handlers/storage/global-this-storage.spec.ts
@@ -1,70 +1,95 @@
+import { RemoteInfo } from '../remote-info';
 import { globalThisStorageEntry } from './global-this-storage';
 import { NfCache, nfNamespace } from './storage.contract';
 
 describe('globalThisStorageEntry', () => {
-    // type MOCK_CACHE = NfCache & {key1: string, key2: string}
+
+    const MOCK_REMOTE_INFO = (): RemoteInfo => ({
+        scopeUrl: "http://localhost:3001/",
+        remoteName: "team/mfe1",
+        exposes: [{moduleName: "./comp", url: "http://localhost:3001/comp.js"}]
+    });
+
+    const MOCK_REMOTE_INFO_II = (): RemoteInfo => ({
+        scopeUrl: "http://localhost:3002/",
+        remoteName: "team/mfe2",
+        exposes: [{moduleName: "./comp", url: "http://localhost:3002/comp.js"}]
+    });
 
     beforeEach(() => {
         delete (globalThis as any)[nfNamespace];
     });
 
     it('creates namespace if it does not exist', () => {
-        globalThisStorageEntry('externals', {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"});
+        globalThisStorageEntry('remotes', {"team/mfe1": MOCK_REMOTE_INFO()});
+
         expect((globalThis as any)[nfNamespace]).toEqual({
-            "externals": {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"}
+            "remotes": {"team/mfe1": MOCK_REMOTE_INFO()}
         });
     });
 
     describe('get', () => {
         it('get should return the fallback value', () => {
-            const entry = globalThisStorageEntry('externals', {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"});
-            const expected = entry.get();
+            const entry = globalThisStorageEntry('remotes', {"team/mfe1": MOCK_REMOTE_INFO()});
 
-            expect(expected).toEqual({
-                    "rxjs@7.8.1": "http://localhost:3001/rxjs.js"
-            });
+            const expected = {"team/mfe1": MOCK_REMOTE_INFO()};
+
+            expect(entry.get()).toEqual(expected);
         });
 
         it('not allow any mutations', () => {
-            const entry = globalThisStorageEntry('externals', {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"});
-            const keyA = entry.get();
-            keyA["tslib@2.8.1"] = "http://localhost:3001/tslib.js";
-             
-            const expected = entry.get();
+            const entry = globalThisStorageEntry('remotes', {"team/mfe1": MOCK_REMOTE_INFO()});
 
-            expect(expected).toEqual({
-                    "rxjs@7.8.1": "http://localhost:3001/rxjs.js"
-            });
+            const expected = {"team/mfe1": MOCK_REMOTE_INFO()};
+
+            const keyA = entry.get();
+            keyA["team/mfe1"] = MOCK_REMOTE_INFO_II();
+
+            expect(entry.get()).toEqual(expected);
         });
     })
 
     describe('set', () => {
         it('set stores value in globalThis namespace', () => {
-            const entry = globalThisStorageEntry('externals', {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"});
-            entry.set({"tslib@2.8.1": "http://localhost:3001/tslib.js"});
+            const entry = globalThisStorageEntry('remotes', {"team/mfe1": MOCK_REMOTE_INFO()});
+            const expected = {"team/mfe2": MOCK_REMOTE_INFO_II()};
 
-            expect(entry.get()).toEqual({"tslib@2.8.1": "http://localhost:3001/tslib.js"});
+            entry.set({"team/mfe2": MOCK_REMOTE_INFO_II()});
+
+            expect(entry.get()).toEqual(expected);
         });
 
         it('maintains separate values for different keys', () => {
-            const entry1 = globalThisStorageEntry('externals', {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"});
-            const entry2 = globalThisStorageEntry('baseUrlToRemoteNames', { "http://localhost:3001": "team/mfe1" });
+            const entry1 = globalThisStorageEntry('externals', {
+                global:{
+                    "rxjs": {version: "7.8.1", requiredVersion: "~7.8.0", url: "http://localhost:3001/rxjs.js"}
+                }
+            });
+            const entry2 = globalThisStorageEntry('remotes', {"team/mfe1": MOCK_REMOTE_INFO()});
             
-            entry1.set({"tslib@2.8.1": "http://localhost:3001/tslib.js"});
-            entry2.set({ "http://localhost:3002": "team/mfe2" });
+            entry1.set({
+                global:{
+                    "tslib": {version: "2.8.1", requiredVersion: "~2.8.0", url: "http://localhost:3001/tslib.js"}
+                }
+            });
+            entry2.set({ "team/mfe2": MOCK_REMOTE_INFO_II() });
             
-            expect(entry1.get()).toEqual({"tslib@2.8.1": "http://localhost:3001/tslib.js"});
-            expect(entry2.get()).toEqual({ "http://localhost:3002": "team/mfe2" });
+            expect(entry1.get()).toEqual({
+                global:{
+                    "tslib": {version: "2.8.1", requiredVersion: "~2.8.0", url: "http://localhost:3001/tslib.js"}
+                }
+            });
+            expect(entry2.get()).toEqual({ "team/mfe2": MOCK_REMOTE_INFO_II() });
         });
 
         it('not allow any mutations', () => {
-            const entry = globalThisStorageEntry('externals', {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"});
-            const newEntry: Record<string, string> = {"tslib@2.8.1": "http://localhost:3001/tslib.js"};
+            const entry = globalThisStorageEntry('remotes', {"team/mfe1": MOCK_REMOTE_INFO()});
+            const newEntry = {"team/mfe2": MOCK_REMOTE_INFO_II()} as any;
             entry.set(newEntry);
 
-            newEntry["MALICOUS_INJECT"] = "http://localhost:3005/script.js";
+            newEntry["MALICOUS_INJECT"] = "BAD_SCRIPT.js";
 
-            expect(entry.get()).toEqual({"tslib@2.8.1": "http://localhost:3001/tslib.js"});
+            expect(entry.get()).toEqual({"team/mfe2": MOCK_REMOTE_INFO_II()});
         });
     });