+++ src/lib/handlers/storage/storage.handler.spec.ts
@@ -1,50 +1,48 @@
 import { mockStorageEntry } from "../../../mock/storage.mock";
+import { RemoteInfo } from "../remote-info";
 import { NfCache, StorageHandler, StorageOf } from "./storage.contract";
 import { storageHandlerFactory } from "./storage.handler";
-import { Remote } from "../remote-info/remote-info.contract";
-import { SharedInfo } from "@softarc/native-federation-runtime";
 
 describe('storageHandler', () => {
 
     let storage: NfCache;
     let storageHandler: StorageHandler<NfCache>;
 
-    const REMOTE_MFE1_MOCK: () => Remote = () => 
-        JSON.parse(JSON.stringify({
-            name: 'team/mfe1', 
-            shared: [
-                {
-                    packageName: "rxjs",
-                    outFileName: "rxjs.js",
-                    requiredVersion: "~7.8.0",
-                    singleton: true,
-                    strictVersion: true,
-                    version: "7.8.1",
-                }
-            ] as SharedInfo[], 
-            exposes: [{key: './comp', outFileName: 'comp.js'}], 
-            baseUrl: 'http://localhost:3001'
-        }))
+    const MOCK_REMOTE_INFO = (): RemoteInfo => ({
+        scopeUrl: "http://localhost:3001/",
+        remoteName: "team/mfe1",
+        exposes: [{moduleName: "./comp", url: "http://localhost:3001/comp.js"}]
+    });
+
+    const MOCK_REMOTE_INFO_II = (): RemoteInfo => ({
+        scopeUrl: "http://localhost:3002/",
+        remoteName: "team/mfe2",
+        exposes: [{moduleName: "./comp", url: "http://localhost:3002/comp.js"}]
+    });
+
+    const MOCK_REMOTE_INFO_III = (): RemoteInfo => ({
+        scopeUrl: "http://localhost:3003/",
+        remoteName: "team/mfe3",
+        exposes: [{moduleName: "./comp", url: "http://localhost:3003/comp.js"}]
+    });
 
     beforeEach(() => {
         storage = {
-            externals: {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"},
-            remoteNamesToRemote: { "team/mfe1": REMOTE_MFE1_MOCK() },
-            baseUrlToRemoteNames: { "http://localhost:3001": "team/mfe1" }
+            externals: {
+                global: {"rxjs": {version: "7.8.1", requiredVersion: "~7.8.0", url: "http://localhost:3001/rxjs.js"} }
+            },
+            remotes: { "team/mfe1": MOCK_REMOTE_INFO() },
         };
-        storageHandler = storageHandlerFactory(storage, mockStorageEntry);
+        storageHandler = storageHandlerFactory({cache: storage, toStorageEntry: mockStorageEntry});
     });
 
     describe('entry', () => {
         it('should get the initial values', () => {            
             expect(storageHandler.entry("externals").get()).toEqual(
-                {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"}
+                {global: {"rxjs": {version: "7.8.1", requiredVersion: "~7.8.0", url:"http://localhost:3001/rxjs.js"}}}
             );
-            expect(storageHandler.entry("remoteNamesToRemote").get()).toEqual(
-                { "team/mfe1": REMOTE_MFE1_MOCK() }
-            );
-            expect(storageHandler.entry("baseUrlToRemoteNames").get()).toEqual(
-                { "http://localhost:3001": "team/mfe1" }
+            expect(storageHandler.entry("remotes").get()).toEqual(
+                { "team/mfe1": MOCK_REMOTE_INFO() }
             );
         });
     });
@@ -52,20 +50,18 @@ describe('storageHandler', () => {
     describe('fetch', () => {
         it('should get the value', () => {
             expect(storageHandler.fetch("externals")).toEqual(
-                {"rxjs@7.8.1": "http://localhost:3001/rxjs.js"}
+                {global: {"rxjs": {version: "7.8.1", requiredVersion: "~7.8.0", url:"http://localhost:3001/rxjs.js"}}}
             );
-            expect(storageHandler.fetch("remoteNamesToRemote")).toEqual(
-                { "team/mfe1": REMOTE_MFE1_MOCK() }
-            );
-            expect(storageHandler.fetch("baseUrlToRemoteNames")).toEqual(
-                { "http://localhost:3001": "team/mfe1" }
+            expect(storageHandler.fetch("remotes")).toEqual(
+                { "team/mfe1": MOCK_REMOTE_INFO() }
             );
         });
     })
 
     describe('update', () => {
         it('should update the value', () => {
-            const expected: Record<string,string> = {"tslib@2.8.1": "http://localhost:3001/tslib.js"};
+            const expected = {global: {"tslib": {version: "2.8.1", requiredVersion: "~2.8.0", url:"http://localhost:3001/tslib.js"}}}
+
             storageHandler.update("externals", _ => (expected));
             
             const actual = storageHandler.fetch("externals");
@@ -75,37 +71,40 @@ describe('storageHandler', () => {
 
         it('should update entries multiple times', () => {
             storageHandler
-                .update("baseUrlToRemoteNames", _ => ({"http://localhost:3002": "team/mfe2"}))
-                .update("baseUrlToRemoteNames", _ => ({"http://localhost:3003": "team/mfe3"}));
+                .update("remotes", _ => ({"team/mfe2": MOCK_REMOTE_INFO_II()}))
+                .update("remotes", _ => ({"team/mfe3": MOCK_REMOTE_INFO_III()}))
             
-            const actual = storageHandler.fetch("baseUrlToRemoteNames");
+            const actual = storageHandler.fetch("remotes");
 
-            expect(actual).toEqual({"http://localhost:3003": "team/mfe3"});
+            expect(actual).toEqual({"team/mfe3": MOCK_REMOTE_INFO_III()});
         });
 
         it('should update multiple entries', () => {
             storageHandler
-                .update("externals", _ => ({"tslib@2.8.1": "http://localhost:3001/tslib.js"}))
-                .update("baseUrlToRemoteNames", _ => ({"http://localhost:3002": "team/mfe2"}));
+                .update("externals", _ => ({global: {"tslib": {version: "2.8.1", requiredVersion: "~2.8.0", url:"http://localhost:3001/tslib.js"}}}))
+                .update("remotes", _ => ({"team/mfe2": MOCK_REMOTE_INFO_II()}))
             
             const actual1 = storageHandler.fetch("externals");
-            expect(actual1).toEqual({"tslib@2.8.1": "http://localhost:3001/tslib.js"});
+            expect(actual1).toEqual({global: {"tslib": {version: "2.8.1", requiredVersion: "~2.8.0", url:"http://localhost:3001/tslib.js"}}});
 
-            const actual2 = storageHandler.fetch("baseUrlToRemoteNames");
-            expect(actual2).toEqual({"http://localhost:3002": "team/mfe2"});
+            const actual2 = storageHandler.fetch("remotes");
+            expect(actual2).toEqual({"team/mfe2": MOCK_REMOTE_INFO_II()});
         });
 
         it('should be able to manipulate the current entry value', () => {
             storageHandler
                 .update("externals", v => {
-                    v["tslib@2.8.1"] = "http://localhost:3001/tslib.js";
+                    v["global"]["tslib"] = {version: "2.8.1", requiredVersion: "~2.8.0", url:"http://localhost:3001/tslib.js"};
                     return v;
                 })
             
             const actual = storageHandler.fetch("externals");
             expect(actual).toEqual({
-                "rxjs@7.8.1": "http://localhost:3001/rxjs.js",
-                "tslib@2.8.1": "http://localhost:3001/tslib.js"
+                global: {
+                "rxjs": {version: "7.8.1", requiredVersion: "~7.8.0", url:"http://localhost:3001/rxjs.js"},
+                "tslib": {version: "2.8.1", requiredVersion: "~2.8.0", url:"http://localhost:3001/tslib.js"}
+                }
+
             });
         });
     });
@@ -114,10 +113,8 @@ describe('storageHandler', () => {
         it('should get the storage object', () => {
             const actual: StorageOf<NfCache> = storageHandler.get();
 
-            expect(actual["externals"].get()).toEqual({"rxjs@7.8.1": "http://localhost:3001/rxjs.js"});
-            expect(actual["remoteNamesToRemote"].get()).toEqual({"team/mfe1": REMOTE_MFE1_MOCK()});
-            expect(actual["baseUrlToRemoteNames"].get()).toEqual({"http://localhost:3001": "team/mfe1"});
-
+            expect(actual["externals"].get()).toEqual({global: {"rxjs": {version: "7.8.1", requiredVersion: "~7.8.0", url:"http://localhost:3001/rxjs.js"}}});
+            expect(actual["remotes"].get()).toEqual({"team/mfe1": MOCK_REMOTE_INFO()});
         });
     });