+++ src/lib/handlers/storage/storage.handler.ts
@@ -1,18 +1,20 @@
-import type { StorageHandler, StorageEntry, NfCache, StorageEntryCreator, StorageOf } from "./storage.contract";
+import type { StorageHandler, StorageEntry, NfCache, StorageOf } from "./storage.contract";
+import type { StorageConfig } from "../../utils/config/config.contract";
 
-function storageHandlerFactory<TCache extends NfCache>(
-    cache: TCache,
-    cacheEntryCreator: StorageEntryCreator
-): StorageHandler<TCache> {
+const storageHandlerFactory = <TCache extends NfCache>(
+    {cache, toStorageEntry}: StorageConfig<TCache>
+): StorageHandler<TCache>=> {
 
     const STORAGE: StorageOf<TCache> = (Object.entries(cache) as { [K in keyof TCache]: [K, TCache[K]]; }[keyof TCache][])
         .reduce(
             (acc, [key, value]) => ({
                 ...acc,
-                [key]: cacheEntryCreator<TCache, typeof key>(key, value)
+                [key]: toStorageEntry<TCache, typeof key>(key, value)
             }),
             {} as StorageOf<TCache>
         );
+
+    
 
     function entry<K extends keyof TCache>(key: K): StorageEntry<TCache[K]> {
         return STORAGE[key];
@@ -23,20 +25,21 @@ function storageHandlerFactory<TCache extends NfCache>(
     };
 
     function update<K extends keyof TCache>(
-        this:StorageHandler<TCache>,
         key: K,
         updateFn: (v: TCache[K]) => TCache[K]
     ): StorageHandler<TCache> {
         const newVal = updateFn(fetch(key));
         STORAGE[key].set(newVal);
-        return this;
+        return handler;
     };
 
     function get(): StorageOf<TCache>{
         return STORAGE;
     }
 
-    return {fetch, get, entry, update};
+    const handler = {fetch, get, entry, update};
+
+    return handler;
 }
 
 export {storageHandlerFactory};